+++ meson.build
@@ -31,9 +31,9 @@ files = [
   'src/dict.c',
   'src/disk.c',
   'src/fmt.c',
-  'src/ht.c',
-  'src/ht_p.c',
-  'src/ht_u.c',
+  'src/ht_uu.c',
+  'src/ht_pp.c',
+  'src/ht_up.c',
   'src/journal.c',
   'src/json.c',
   #'src/json/api.c',
