+++ src/ht_pp.c
@@ -1,14 +1,15 @@
-#include "ht_u.h"
+#include "ht_pp.h"
+#include "sdb.h"
 
 #include "ht_inc.c"
 
 static HtName_(Ht)* internal_ht_default_new(ut32 size, ut32 prime_idx, HT_(DupValue) valdup, HT_(KvFreeFunc) pair_free, HT_(CalcSizeV) calcsizeV) {
 	HT_(Options) opt = {
-		.cmp = NULL,
-		.hashfn = NULL, // TODO: use a better hash function for numbers
-		.dupkey = NULL,
+		.cmp = (HT_(ListComparator))strcmp,
+		.hashfn = (HT_(HashFunction))sdb_hash,
+		.dupkey = (HT_(DupKey))strdup,
 		.dupvalue = valdup,
-		.calcsizeK = NULL,
+		.calcsizeK = (HT_(CalcSizeK))strlen,
 		.calcsizeV = calcsizeV,
 		.freefn = pair_free,
 		.elem_size = sizeof (HT_(Kv)),
@@ -16,13 +17,18 @@ static HtName_(Ht)* internal_ht_default_new(ut32 size, ut32 prime_idx, HT_(DupVa
 	return internal_ht_new (size, prime_idx, &opt);
 }
 
+// creates a default HtPP that has strings as keys
 SDB_API HtName_(Ht)* Ht_(new)(HT_(DupValue) valdup, HT_(KvFreeFunc) pair_free, HT_(CalcSizeV) calcsizeV) {
 	return internal_ht_default_new (ht_primes_sizes[0], 0, valdup, pair_free, calcsizeV);
 }
 
-// creates a default HtUP that does not dup, nor free the values
+static void free_kv_key(HT_(Kv) *kv) {
+	free (kv->key);
+}
+
+// creates a default HtPP that has strings as keys but does not dup, nor free the values
 SDB_API HtName_(Ht)* Ht_(new0)(void) {
-	return Ht_(new) (NULL, NULL, NULL);
+	return Ht_(new) (NULL, free_kv_key, NULL);
 }
 
 SDB_API HtName_(Ht)* Ht_(new_size)(ut32 initial_size, HT_(DupValue) valdup, HT_(KvFreeFunc) pair_free, HT_(CalcSizeV) calcsizeV) {
