+++ src/ht_pp.h
