+++ src/ht_uu.c
@@ -1,3 +1,3 @@
-#include "ht.h"
+#include "ht_uu.h"
 
 #include "ht_inc.c"
