+++ src/sdbht.h
@@ -1,7 +1,7 @@
 #ifndef __SDB_HT_H
 #define __SDB_HT_H
 
-#include "ht_p.h"
+#include "ht_pp.h"
 
 /** keyvalue pair **/
 typedef struct sdb_kv {
