+++ test/unit/test_hash.c
@@ -1,6 +1,7 @@
 #include "minunit.h"
 #include <sdb.h>
-#include <ht_u.h>
+#include <ht_up.h>
+#include <ht_pp.h>
 
 typedef struct _test_struct {
 	char *name;
